Created a helper function for showCategories
